Split the scene_parser module into multiple files.
